+++ includes/input.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   input.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: user42 <user42@student.42.fr>              +#+  +:+       +#+        */
+/*   By: lcouto <lcouto@student.42sp.org.br>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/28 20:31:27 by lfrasson          #+#    #+#             */
-/*   Updated: 2021/07/15 01:35:25 by user42           ###   ########.fr       */
+/*   Updated: 2021/07/31 22:49:17 by lcouto           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -57,7 +57,7 @@
 # define ESC_BOLD_WHITE "\001\033[1;37m\002"
 # define ESC_RESET_COLOR "\001\033[0m\002"
 
-char	*create_prompt(void);
+char	*create_prompt(bool *emoji);
 void	define_input_signals(void);
 void	define_exec_signals(void);