+++ sources/system/prompt.c
@@ -6,7 +6,7 @@
 /*   By: lcouto <lcouto@student.42sp.org.br>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/28 20:40:47 by lfrasson          #+#    #+#             */
-/*   Updated: 2021/07/26 19:43:31 by lcouto           ###   ########.fr       */
+/*   Updated: 2021/07/31 22:52:25 by lcouto           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,23 +18,6 @@ static char	*get_pwd(void)
 
 	buffer = NULL;
 	return (getcwd(buffer, 0));
-}
-
-static void	free_prompt_strings(char *name, char *pwd, char *end)
-{
-	free(name);
-	free(pwd);
-	free(end);
-}
-
-static int	get_prompt_len(char *name, char *pwd, char *end)
-{
-	int	len;
-
-	len = ft_strlen(pwd);
-	len += ft_strlen(name);
-	len += ft_strlen(end);
-	return (len);
 }
 
 static char	*concatenate_prompt(char *name, char *pwd, char *end)
@@ -50,23 +33,51 @@ static char	*concatenate_prompt(char *name, char *pwd, char *end)
 			ESC_BOLD_RED,
 			end,
 			ESC_RESET_COLOR);
-	free_prompt_strings(name, pwd, end);
+	free(name);
+	free(pwd);
+	free(end);
 	return (prompt);
 }
 
-char	*create_prompt(void)
+static void	display_welcome_message(void)
+{
+	char	*path;
+	char	*local_paths;
+	char	**cmd;
+
+	path = (ft_strdup(hashmap_search(g_minishell.env, "PATH")));
+	local_paths = variadic_strjoin(7, path, ":",
+			hashmap_search(g_minishell.env, "PWD"), ":",
+			hashmap_search(g_minishell.env, "PWD"), "/sources/system", ":.");
+	cmd = ft_split("welcome", ' ');
+	hashmap_insert("PATH", local_paths, g_minishell.env);
+	execute(cmd);
+	free(path);
+	free(local_paths);
+	free_2d_array(cmd);
+}
+
+static char	*create_name(bool *emoji)
+{
+	if (!*emoji)
+	{
+		*emoji = true;
+		display_welcome_message();
+		return (ft_strdup("\n"));
+	}
+	if (g_minishell.error_status == 0)
+		return (ft_strdup("😈🔥 MINIHELL 🔥😈:"));
+	return (ft_strdup("👿🔥 MINIHELL 🔥👿:"));
+}
+
+char	*create_prompt(bool *emoji)
 {
 	char	*name;
 	char	*pwd;
 	char	*end;
-	char	prompt_len;
 
 	pwd = get_pwd();
-	if (g_minishell.error_status == 0)
-		name = ft_strdup("😈🔥 MINIHELL 🔥😈:");
-	else
-		name = ft_strdup("👿🔥 MINIHELL 🔥👿:");
+	name = create_name(emoji);
 	end = ft_strdup("$ ");
-	prompt_len = get_prompt_len(name, pwd, end);
 	return (concatenate_prompt(name, pwd, end));
 }