+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
+## - 2021-07-31
+### Added
+- Added prompt welcome message.
+
 ## - 2021-07-28
 ### Added
 - Added file here document redirection feature.